requirements.txt
# This file was generated by pipenv based on the Pipfile. DO NOT EDIT MANUALLY!
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9; sys_platform == 'win32'
colorlog==3.1.4
idna==2.7
pymongo==3.7.0
pyyaml==3.13
raven==6.9.0
requests==2.19.1
urllib3==1.23; python_version != '3.0.*'
